requirements.txt
Flask>=2.0.1
transformers>=4.7.0
googlesearch-python>=1.0.1
beautifulsoup4>=4.9.3
pandas>=1.3.3
requests>=2.26.0
TensorFlow >= 2.0